+++ particle_sim/src/serial/SPH.cpp
@@ -3,30 +3,30 @@
 #include <algorithm>
 
 // defining the kernel constants
-const double SPH::K_POLY6 = 315.0 / (64.0 * (double)M_PI * std::pow(H, 9.0));
-const double SPH::K_SPIKY = -45.0 / ((double)M_PI * std::pow(H, 6.0));
-const double SPH::K_VISC  =  45.0 / ((double)M_PI * std::pow(H, 6.0));
+const float SPH::K_POLY6 = 315.0f / (64.0f * (float)M_PI * std::pow(H, 9.0f));
+const float SPH::K_SPIKY = -45.0f / ((float)M_PI * std::pow(H, 6.0f));
+const float SPH::K_VISC  =  45.0f / ((float)M_PI * std::pow(H, 6.0f));
 
 // *** Kernel Functions ***
 // from the STAR paper
 
 // good for density accumulation
-double SPH::W_poly6(double r2) const {
-    if (r2 >= H2) return 0.0;
-    double d = H2 - r2;
+float SPH::W_poly6(float r2) const {
+    if (r2 >= H2) return 0.0f;
+    float d = H2 - r2;
     return K_POLY6 * d * d * d;
 }
 
 // good for a repulsive contribution in the pressure force sum
-Vec3 SPH::gradW_spiky(const Vec3& rij, double r) const {
+Vec3 SPH::gradW_spiky(const Vec3& rij, float r) const {
     if (r >= H || r < 1e-6f) return {};
-    double d = H - r;
-    double coef = K_SPIKY * d * d / r;
+    float d = H - r;
+    float coef = K_SPIKY * d * d / r;
     return rij * coef;
 }
 
-double SPH::lapW_visc(double r) const {
-    if (r >= H) return 0.0;
+float SPH::lapW_visc(float r) const {
+    if (r >= H) return 0.0f;
     return K_VISC * (H - r);
 }
 
@@ -38,15 +38,15 @@ SPH::SPH() : grid(H) {
 
 void SPH::reset() {
     particles.clear();
-    for (double x = (0.5 * BMIN); x < (0.5 * BMAX); x += SPACING) { // left half of the box
-        for (double y = BMIN + SPACING; y < (0.5 * BMAX); y += SPACING) { // full height
-            for (double z = BMIN + SPACING; z < BMAX - SPACING; z += SPACING) { // full depth
+    for (float x = (0.5f * BMIN); x < (0.5f * BMAX); x += SPACING) { // left half of the box
+        for (float y = BMIN + SPACING; y < (0.5f * BMAX); y += SPACING) { // full height
+            for (float z = BMIN + SPACING; z < BMAX - SPACING; z += SPACING) { // full depth
                 Particle p;
                 p.pos = {x, y, z};
-                p.vel = {0.0, 0.0, 0.0};
-                p.acc = {0.0, 0.0, 0.0};
+                p.vel = {0.0f, 0.0f, 0.0f};
+                p.acc = {0.0f, 0.0f, 0.0f};
                 p.density  = RHO0;
-                p.pressure = 0.0;
+                p.pressure = 0.0f;
                 particles.push_back(p);
             }
         }
@@ -65,7 +65,7 @@ void SPH::step() {
 // initalized the spacial hash grid with particles (and their positions)
 void SPH::buildGrid() {
     grid.clear();
-    for (int i = 0; i < particles.size(); ++i) grid.insert(i, particles[i].pos);
+    for (int i = 0; i < (int)particles.size(); ++i) grid.insert(i, particles[i].pos);
 }
 
 // computes both pressure and density for all particles using the equations
@@ -75,31 +75,31 @@ void SPH::computeDensityPressure() {
         grid.query(pi.pos, H, nbrs);
 
         // density first
-        pi.density = 0.0;
+        pi.density = 0.0f;
         for (int j : nbrs) {
             Vec3  rij = pi.pos - particles[j].pos;
-            double r2  = rij.length2();
+            float r2  = rij.length2();
             pi.density += MASS * W_poly6(r2);
         }
         // to avoid division by zero or negative pressure
-        pi.density = std::max(pi.density, RHO0 * 0.1);
+        pi.density = std::max(pi.density, RHO0 * 0.1f);
 
         // pressure
         // tait equation from STAR paper
-        double ratio = pi.density / RHO0;
-        double r2t = ratio * ratio;
-        double r4t = r2t * r2t;
-        double r7 = r4t * r2t * ratio;
-        pi.pressure = std::max(0.0, B_PRESS * (r7 - 1.0));
+        float ratio = pi.density / RHO0;
+        float r2t = ratio * ratio;
+        float r4t = r2t * r2t;
+        float r7 = r4t * r2t * ratio;
+        pi.pressure = std::max(0.0f, B_PRESS * (r7 - 1.0f));
     }
 }
 
 // use previously calculated pressure and density along with other items
 // here to calculate forces on particles
 void SPH::computeForces() {
-    const Vec3 gravity{0.0, -GRAVITY, 0.0};
+    const Vec3 gravity{0.0f, -GRAVITY, 0.0f};
 
-    for (int i = 0; i < particles.size(); ++i) {
+    for (int i = 0; i < (int)particles.size(); ++i) {
         auto& pi = particles[i];
         Vec3 a_press{}, a_visc{};
 
@@ -109,16 +109,16 @@ void SPH::computeForces() {
             auto& pj = particles[j];
 
             Vec3  rij = pi.pos - pj.pos;
-            double r   = rij.length();
+            float r   = rij.length();
             if (r >= H || r < 1e-6f) continue;
 
             // acceleration from pressure
-            double pterm = pi.pressure / (pi.density * pi.density)
+            float pterm = pi.pressure / (pi.density * pi.density)
                 + pj.pressure / (pj.density * pj.density);
             a_press += gradW_spiky(rij, r) * (-MASS * pterm);
 
             // acceleration from viscosity (damps velocity)
-            double wlap = lapW_visc(r);
+            float wlap = lapW_visc(r);
             a_visc += (pj.vel - pi.vel) * (MU * MASS / pj.density * wlap);
         }
 
@@ -137,12 +137,12 @@ void SPH::integrate() {
 }
 
 void SPH::enforceBoundary(Particle& p) {
-    auto reflect = [&](double& pos, double& vel, double mn, double mx) {
+    auto reflect = [&](float& pos, float& vel, float mn, float mx) {
         if (pos < mn) {
-            pos = mn + 2e-4f; if (vel < 0.0) vel *= -RESTITUTION;
+            pos = mn + 2e-4f; if (vel < 0.0f) vel *= -RESTITUTION;
         }
         if (pos > mx) {
-            pos = mx - 2e-4f; if (vel > 0.0) vel *= -RESTITUTION;
+            pos = mx - 2e-4f; if (vel > 0.0f) vel *= -RESTITUTION;
         }
     };
     reflect(p.pos.x, p.vel.x, BMIN, BMAX);
