+++ particle_sim/src/serial/SPH.h
@@ -7,8 +7,8 @@ struct Particle {
     Vec3 pos;
     Vec3 vel;
     Vec3 acc;
-    double density;
-    double pressure;
+    float density;
+    float pressure;
 };
 
 // 'Weakly Compressible' SPH simulation.
@@ -20,36 +20,36 @@ struct Particle {
 // I'll refer to this paper as 'the STAR paper' throughout these files
 class SPH {
 public:
-    static constexpr double H = 0.13; // smoothing length used in smoothing kernel
-    static constexpr double H2 = H * H;
+    static constexpr float H = 0.13f; // smoothing length used in smoothing kernel
+    static constexpr float H2 = H * H;
 
-    static constexpr double RHO0 = 1000.0; // rest density
+    static constexpr float RHO0 = 1000.0f; // rest density
     // particle spacing. note that increasing/decreasing
     // this value increases and decreases the number of particles
     // spawned.
-    static constexpr double SPACING = H * 0.75;
-    static constexpr double MASS = RHO0 * SPACING * SPACING * SPACING; // explicit mass
+    static constexpr float SPACING = H * 0.75f;
+    static constexpr float MASS = RHO0 * SPACING * SPACING * SPACING; // explicit mass
 
     // some parameters for the so called 'equation of state (EOS)' calculations
-    static constexpr double GAMMA = 7.;
-    static constexpr double C_SOUND = 30.0;
-    static constexpr double B_PRESS = RHO0 * C_SOUND * C_SOUND / GAMMA;
+    static constexpr float GAMMA = 7.0f;
+    static constexpr float C_SOUND = 30.0f;
+    static constexpr float B_PRESS = RHO0 * C_SOUND * C_SOUND / GAMMA;
 
-    static constexpr double MU = 0.13; // dynamic viscosity (may be increased or decreased as needed)
-    static constexpr double GRAVITY = 9.81;
+    static constexpr float MU = 0.13f; // dynamic viscosity (may be increased or decreased as needed)
+    static constexpr float GRAVITY = 9.81f;
 
     // for time integration
-    static constexpr double DT = 0.00025;
+    static constexpr float DT = 0.00025f;
 
     // some paramters for the axi-alligned bounding box
-    static constexpr double BMIN = -3.0;
-    static constexpr double BMAX =  3.0;
-    static constexpr double RESTITUTION = 0.4; // wall collision damping
+    static constexpr float BMIN = -3.0f;
+    static constexpr float BMAX =  3.0f;
+    static constexpr float RESTITUTION = 0.4f; // wall collision damping
 
     // precomputed kernel constants that are defined in SPH.cpp
-    static const double K_POLY6;
-    static const double K_SPIKY;
-    static const double K_VISC;
+    static const float K_POLY6;
+    static const float K_SPIKY;
+    static const float K_VISC;
 
     std::vector<Particle> particles;
 
@@ -64,9 +64,9 @@ private:
 
     // the kernel functions
     // read the STAR paper for more details on these
-    double W_poly6(double r2) const;
-    Vec3  gradW_spiky(const Vec3& rij, double r) const;
-    double lapW_visc(double r) const;
+    float W_poly6(float r2) const;
+    Vec3  gradW_spiky(const Vec3& rij, float r) const;
+    float lapW_visc(float r) const;
 
     // simulation passes called in step()
     void buildGrid();
